img_api/pyproject.toml
[tool.poetry]
name = "img-api"
version = "0.1.0"
description = ""
authors = ["kooznitsa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
pydantic-settings = "^2.3.3"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
aiohttp = "^3.9.5"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.7"
sqlmodel = "^0.0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
